+++ src/components/Mui/MuiButton.stories.js
@@ -1,6 +1,6 @@
 import Button from '@mui/material/Button';
 
-import { MuiProvider } from '../../styles/MuiProvider';
+import { MuiProvider } from '../../styles';
 
 export default {
   title: 'Mui/Button',
