+++ src/components/Mui/MuiChip.stories.js
@@ -1,6 +1,6 @@
 import Chip from '@mui/material/Chip';
 
-import { MuiProvider } from '../../styles/MuiProvider';
+import { MuiProvider } from '../../styles';
 
 export default {
   title: 'Mui/Chip',
