+++ src/components/Mui/MuiIconButton.stories.js
@@ -1,7 +1,7 @@
 import IconButton from '@mui/material/IconButton';
 import SettingsIcon from '@mui/icons-material/Settings';
 
-import { MuiProvider } from '../../styles/MuiProvider';
+import { MuiProvider } from '../../styles';
 
 export default {
   title: 'Mui/IconButton',
